Clean up code after rebase
