+++ src/platform/vita/vita.c
@@ -5,9 +5,8 @@
 #include "graphics/screen.h"
 #include "input/mouse.h"
 #include "input/touch.h"
-
-#include <vita2d.h>
-#include <vitasdk.h>
+#include "input/keyboard.h"
+#include "platform/vita/vita_keyboard.h"
 
 #include <string.h>
 #include <stdio.h>
@@ -29,6 +28,12 @@ static void center_mouse_cursor(void)
     mouse_set_position(x, y);
 }
 
+static struct {
+    uint8_t *text;
+    int requested;
+    int max_length;
+} vkbd;
+
 void platform_init_callback(void)
 {
     touch_set_mode(TOUCH_MODE_TOUCHPAD);
@@ -48,3 +53,28 @@ const char *vita_prepend_path(const char *path)
     strncpy(prepended_path + PREPEND_PATH_OFFSET, path, 2 * FILE_NAME_MAX * sizeof(char) - PREPEND_PATH_OFFSET - 1);
     return prepended_path;
 }
+
+static void vita_start_text_input(void)
+{
+    if (!keyboard_is_capturing()) {
+        return;
+    }
+    const uint8_t *text = vita_keyboard_get(keyboard_get_text(), keyboard_get_max_text_length());
+    keyboard_set_text(text);
+}
+
+void platform_per_frame_callback(void)
+{
+    if (vkbd.requested) {
+        vita_start_text_input();
+        vkbd.requested = 0;
+    }
+}
+
+void platform_show_virtual_keyboard(void)
+{
+    vkbd.requested = 1;
+}
+
+void platform_hide_virtual_keyboard(void)
+{}
